SentenceDisplayer: draws the recorded notes smoothly, independent of frame-rate and beat granularity of the song
PlayerScoreController: do not show bad SentenceRating if there are no scorable notes (e.g. only freestyle notes)

+++ UltraStar Play/Assets/Scenes/Sing/PlayerController/RecordedNote.cs
@@ -8,12 +8,12 @@ public class RecordedNote
     public int RecordedMidiNote { get; set; }
     public int RoundedMidiNote { get; set; }
 
-    public double StartBeat { get; set; }
-    public double EndBeat { get; set; }
+    public int StartBeat { get; set; }
+    public int EndBeat { get; set; }
 
     public Note TargetNote { get; set; }
 
-    public RecordedNote(int recordedMidiNote, int roundedMidiNote, double startBeat, double endBeat, Note targetNote)
+    public RecordedNote(int recordedMidiNote, int roundedMidiNote, int startBeat, int endBeat, Note targetNote)
     {
         RecordedMidiNote = recordedMidiNote;
         RoundedMidiNote = roundedMidiNote;
